+++ src/Relationships/AbstractRelationship.php
@@ -6,13 +6,8 @@ use BadMethodCallException;
 use IlluminateAgnostic\Str\Support\Str;
 use Somnambulist\Collection\Contracts\Collection;
 use Somnambulist\Components\ApiClient\AbstractModel;
-use Somnambulist\Components\ApiClient\Client\Behaviours\DecodeResponseArray;
-use Somnambulist\Components\ApiClient\Client\Contracts\ExpressionInterface;
 use Somnambulist\Components\ApiClient\Client\Query\Expression\CompositeExpression;
-use Somnambulist\Components\ApiClient\Client\Query\Expression\ExpressionBuilder;
-use Somnambulist\Components\ApiClient\Contracts\RelatableInterface;
 use Somnambulist\Components\ApiClient\ModelBuilder;
-use function method_exists;
 use function sprintf;
 
 /**
@@ -21,18 +16,11 @@ use function sprintf;
  * @package    Somnambulist\Components\ApiClient\Relationships
  * @subpackage Somnambulist\Components\ApiClient\Relationships\AbstractRelationship
  *
- * @method ExpressionBuilder expr()
- *
  * @method AbstractRelationship with(...$relationship)
- * @method AbstractRelationship where(ExpressionInterface ...$predicate)
- * @method AbstractRelationship andWhere(ExpressionInterface ...$predicates)
- * @method AbstractRelationship orWhere(ExpressionInterface ...$predicates)
- * @method AbstractRelationship orderBy(string $field, string $dir = 'asc')
- * @method AbstractRelationship addOrderBy(string $field, string $dir = 'asc')
- * @method AbstractRelationship page(int $page = null)
- * @method AbstractRelationship perPage(int $perPage = null)
  * @method AbstractRelationship limit(int $limit = null)
  * @method AbstractRelationship offset(string $offset = null)
+ * @method AbstractRelationship page(int $page = null)
+ * @method AbstractRelationship perPage(int $perPage = null)
  *
  * @method array getWith()
  * @method array getOrderBy()
@@ -44,8 +32,6 @@ use function sprintf;
  */
 abstract class AbstractRelationship
 {
-
-    use DecodeResponseArray;
 
     protected AbstractModel $parent;
     protected AbstractModel $related;
@@ -63,10 +49,12 @@ abstract class AbstractRelationship
 
     public function __call($name, $arguments)
     {
-        if (method_exists($this->query, $name)) {
+        $allowed = ['with', 'limit', 'offset', 'page', 'perPage'];
+
+        if (in_array($name, $allowed) || Str::startsWith($name, 'get')) {
             $ret = $this->query->{$name}(...$arguments);
 
-            if (Str::startsWith($name, 'get') || 'expr' === $name) {
+            if (!$ret instanceof $this->query) {
                 return $ret;
             }