+++ src/Relationships/HasOne.php
@@ -39,7 +39,9 @@ class HasOne extends AbstractRelationship
     {
         $models->each(function (AbstractModel $loaded) use ($relationship) {
             if (null === $data = $loaded->getRawAttribute($this->attributeKey)) {
-                $data = $this->query->with($relationship)->wherePrimaryKey($loaded->getPrimaryKey())->fetchRaw();
+                $data = $this->parent->getResponseDecoder()->object(
+                    $this->query->with($relationship)->wherePrimaryKey($loaded->getPrimaryKey())->fetchRaw()
+                );
 
                 if (isset($data[$this->attributeKey])) {
                     $data = $data[$this->attributeKey];